refactor: merge custom-ordered-editor into custom-ordered-view
